Make normal fits by folder
Former-commit-id: 424a2a7bc8f7d896851c2cbb2970da83341d51ef
Former-commit-id: f77bcd7c06584244ae74494f02231c70cb4942c5
Former-commit-id: 753f2f2b34774af9bc28516c7b61af02dc53ecba
Former-commit-id: 1082062544e315d484fa226a62500320f2a2ae2d
Former-commit-id: 516c5053461dc22360efecab0d766855ad3d9d91 [formerly 3aa020ae4ac6ae2e1766ffbbfbf1bf695e619f58]
Former-commit-id: 71185c4b39c67c1936acaf68ed28f1dad2f4090d

+++ extra_funcs.py
@@ -254,7 +254,7 @@ def dict_to_title(d, N=None, exclude=None):
     else:
         cfg = d
     N0_str = human_format(cfg.N0)
-    title = f"N={N0_str}, β={cfg.beta:.4f}, γ={cfg.gamma:.1f}, σ={cfg.sigma:.1f},  α={cfg.alpha:.1f}, ψ={cfg.psi:.1f}, μ={cfg.mu:.1f}, λ1 = {cfg.Mrate1:.1f}, λ2 = {cfg.Mrate2:.1f}, BB = {cfg.BB}"
+    title = f"N={N0_str}, β={cfg.beta:.4f}, γ={cfg.gamma:.1f}, σ={cfg.sigma:.1f},  α={cfg.alpha:.1f}, ψ={cfg.psi:.1f}, μ={cfg.mu:.1f}, λ1 = {cfg.Mrate1:.1f}, λ2 = {cfg.Mrate2:.1f}, Ninit = {cfg.Ninit}, BB = {cfg.BB}"
     if N:
         title += f", #{N}"
 
@@ -520,22 +520,22 @@ def filename_to_par_string(filename):
     return dict_to_str(filename_to_dotdict(filename))
 
 
-def get_fit_results(filenames, force_rerun=False, num_cores_max=20):
+# def get_fit_results(filenames, force_rerun=False, num_cores_max=20):
 
-    output_filename = 'fit_results.joblib'
-    # out_pickle = output_filename.replace('joblib', 'pkl')
+#     output_filename = 'fit_results.joblib'
+#     # out_pickle = output_filename.replace('joblib', 'pkl')
 
-    if Path(output_filename).exists() and not force_rerun:
-        print("Loading fit results")
-        return joblib.load(output_filename)
-        # return pickle.load(open(out_pickle, "rb" ) )
+#     if Path(output_filename).exists() and not force_rerun:
+#         print("Loading fit results")
+#         return joblib.load(output_filename)
+#         # return pickle.load(open(out_pickle, "rb" ) )
 
-    else:
-        fit_results = calc_fit_results(filenames, num_cores_max=num_cores_max)
-        print(f"Finished fitting, saving results to {output_filename}", flush=True)
-        joblib.dump(fit_results, output_filename)
-        # pickle.dump(fit_results, open(out_pickle, "wb"))
-        return fit_results
+#     else:
+#         fit_results = calc_fit_results(filenames, num_cores_max=num_cores_max)
+#         print(f"Finished fitting, saving results to {output_filename}", flush=True)
+#         joblib.dump(fit_results, output_filename)
+#         # pickle.dump(fit_results, open(out_pickle, "wb"))
+#         return fit_results
 
 
 
@@ -547,34 +547,71 @@ def filenames_to_subgroups(filenames):
     return cfg_str
 
 
+def get_fit_normal_results(filenames, force_rerun=False, num_cores_max=20):
+
+
+    all_fits_file = 'fits_normal.joblib'
+
+    if Path(all_fits_file).exists() and not force_rerun:
+        print("Loading all normal fits", flush=True)
+        return joblib.load(all_fits_file)
+
+    else:
+
+        all_normal_fits = {}
+        cfg_str = filenames_to_subgroups(filenames)
+        print(f"Fitting normal for {len(filenames)} simulations spaced on {len(cfg_str.items())} different runs, please wait.", flush=True)
+
+        for sim_pars, files in tqdm(cfg_str.items()):
+            output_filename = Path('Data/fits_normal') / f'normal_fits_{sim_pars}.joblib'
+            output_filename.parent.mkdir(parents=True, exist_ok=True)
+
+            if output_filename.exists() and not force_rerun:
+                all_normal_fits[sim_pars] = joblib.load(output_filename)
+
+            else:
+                fit_results = calc_fit_results(filenames, num_cores_max=num_cores_max)
+                joblib.dump(fit_results, output_filename)
+                all_normal_fits[sim_pars] = fit_results
+
+        joblib.dump(all_normal_fits, all_fits_file)
+        return all_normal_fits
+
+
+
 def get_fit_Imax_results(filenames, force_rerun=False, num_cores_max=20):
 
 
     bins = np.linspace(0, 1, N_peak_fits+1)
     bin_centers = (bins[1:] + bins[:-1])/2
 
-    all_Imax_fits = {}
+    all_fits_file = 'fits_Imax.joblib'
 
-    cfg_str = filenames_to_subgroups(filenames)
+    if Path(all_fits_file).exists() and not force_rerun:
+        print("Loading all Imax fits", flush=True)
+        all_Imax_fits = joblib.load(all_fits_file)
+        return all_Imax_fits, bin_centers
 
-    print(f"Fitting I_max for {len(filenames)} simulations with spaced on {len(cfg_str.items())} different runs, please wait.", flush=True)
+    else:
 
-    for sim_pars, files in tqdm(cfg_str.items()):
+        all_Imax_fits = {}
+        cfg_str = filenames_to_subgroups(filenames)
+        print(f"Fitting I_max for {len(filenames)} simulations spaced on {len(cfg_str.items())} different runs, please wait.", flush=True)
 
-        output_filename = Path('Data/fits') / f'Imax_fits_{sim_pars}.joblib'
-        output_filename.parent.mkdir(parents=True, exist_ok=True)
+        for sim_pars, files in tqdm(cfg_str.items()):
+            output_filename = Path('Data/fits_Imax') / f'Imax_fits_{sim_pars}.joblib'
+            output_filename.parent.mkdir(parents=True, exist_ok=True)
 
-        if output_filename.exists() and not force_rerun:
-            # print("Loading Imax fit results")
-            all_Imax_fits[sim_pars] = joblib.load(output_filename)
+            if output_filename.exists() and not force_rerun:
+                all_Imax_fits[sim_pars] = joblib.load(output_filename)
 
-        else:
-            fit_results = calc_fit_Imax_results(files, num_cores_max=num_cores_max)
-            # print(f"Finished Imax fitting, saving results to {output_filename}", flush=True)
-            joblib.dump(fit_results, output_filename)
-            all_Imax_fits[sim_pars] = fit_results
+            else:
+                fit_results = calc_fit_Imax_results(files, num_cores_max=num_cores_max)
+                joblib.dump(fit_results, output_filename)
+                all_Imax_fits[sim_pars] = fit_results
 
-    return all_Imax_fits, bin_centers
+        joblib.dump(all_Imax_fits, all_fits_file)
+        return all_Imax_fits, bin_centers
 
 
 def cut_percentiles(x, p1, p2=None):
